update bus to use vector and to operate on bytes

+++ mips_emu/ibus.h
@@ -6,5 +6,5 @@ class IBus
 public:
 	virtual uint32_t Size() = 0;
 	virtual uint32_t Read(uint32_t addr) = 0;
-	virtual void Write(uint32_t addr, uint32_t data) = 0;
+	virtual void Write(uint32_t addr, uint8_t data) = 0;
 };
